requirements.txt
selenium
beautifulsoup4
instagrapi
pillow
python-dotenv
schedule
certifi
requests